requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.0
PyPDF2>=2.0.0
openai>=1.0.0
rouge-score>=0.1.2
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
requests>=2.26.0
nltk>=3.8.0